Move Help Functions to their own file so we can keep logic clean -S
