+++ supervisord.conf
@@ -70,7 +70,7 @@ stderr_logfile=/dev/stderr
 # Go starter snake
 [program:go-starter-snake]
 directory=./snakes/go/starter-snake/
-command=/usr/local/go/bin/go run main.go logic.go
+command=/usr/local/go/bin/go run main.go logic.go helpers.go
 stdout_events_enabled=true
 stderr_events_enabled=true
 stdout_logfile_maxbytes=0
